Fix #2262, separate dispatcher for messages
Isolate the message verification and dispatch from the general message
processing.  Functions in the "task" source file now strictly handle the
command content, and do not get involved in general validation.

Also note: this is mostly a simple code move but does add a "const"
qualifier to the task pipe function where it was missing before.  All of
the handlers were already "const".
