+++ modules/es/fsw/src/cfe_es_dispatch.h
@@ -19,30 +19,26 @@
 /**
  * @file
  *
- * Encapsulates all EVS module internal header files, as well
- * as the public API from all other CFE core modules, OSAL, and PSP.
+ *  cFE Executive Services (ES) dispatch header file
  *
- * This simplifies the set of include files that need to be put at the
- * start of every source file.
  */
 
-#ifndef CFE_EVS_MODULE_ALL_H
-#define CFE_EVS_MODULE_ALL_H
+#ifndef CFE_ES_DISPATCH_H
+#define CFE_ES_DISPATCH_H
 
-/********************* Include Files  ************************/
+/*
+** Includes
+*/
+#include "common_types.h"
 
-#include "cfe.h" /* All CFE+OSAL public API definitions */
-#include "cfe_platform_cfg.h"
+#include "cfe_es_api_typedefs.h"
+#include "cfe_sb_api_typedefs.h"
+#include "cfe_msg_api_typedefs.h"
 
-#include "cfe_msgids.h"
-#include "cfe_perfids.h"
+/*---------------------------------------------------------------------------------------*/
+/**
+ * Reads and processes messages from the executive services command pipe
+ */
+void CFE_ES_TaskPipe(const CFE_SB_Buffer_t *SBBufPtr);
 
-#include "cfe_evs_core_internal.h"
-
-#include "cfe_evs_events.h" /* EVS event IDs */
-#include "cfe_evs_task.h"   /* EVS internal definitions */
-#include "cfe_evs_log.h"    /* EVS log file definitions */
-#include "cfe_evs_utils.h"  /* EVS utility function definitions */
-#include "cfe_evs_dispatch.h"
-
-#endif /* CFE_EVS_MODULE_ALL_H */
+#endif /* CFE_ES_DISPATCH_H */
