+++ modules/es/fsw/src/cfe_es_module_all.h
@@ -49,6 +49,7 @@
 #include "cfe_es_events.h"
 #include "cfe_es_start.h"
 #include "cfe_es_task.h"
+#include "cfe_es_dispatch.h"
 #include "cfe_es_resource.h"
 #include "cfe_es_log.h"
