+++ modules/sb/fsw/src/cfe_sb_dispatch.h
@@ -19,30 +19,29 @@
 /**
  * @file
  *
- * Encapsulates all EVS module internal header files, as well
- * as the public API from all other CFE core modules, OSAL, and PSP.
+ * Purpose:
+ *      This header file contains prototypes for private functions and type
+ *      definitions for SB internal use.
  *
- * This simplifies the set of include files that need to be put at the
- * start of every source file.
+ * Author:   R.McGraw/SSI
+ *
  */
 
-#ifndef CFE_EVS_MODULE_ALL_H
-#define CFE_EVS_MODULE_ALL_H
+#ifndef CFE_SB_DISPATCH_H
+#define CFE_SB_DISPATCH_H
 
-/********************* Include Files  ************************/
+/*
+** Includes
+*/
+#include "common_types.h"
+#include "cfe_sb_api_typedefs.h"
 
-#include "cfe.h" /* All CFE+OSAL public API definitions */
-#include "cfe_platform_cfg.h"
+/*---------------------------------------------------------------------------------------*/
+/**
+ * Processes a single message buffer that has been received from the command pipe
+ *
+ * @param SBBufPtr Software bus buffer pointer
+ */
+void CFE_SB_ProcessCmdPipePkt(const CFE_SB_Buffer_t *SBBufPtr);
 
-#include "cfe_msgids.h"
-#include "cfe_perfids.h"
-
-#include "cfe_evs_core_internal.h"
-
-#include "cfe_evs_events.h" /* EVS event IDs */
-#include "cfe_evs_task.h"   /* EVS internal definitions */
-#include "cfe_evs_log.h"    /* EVS log file definitions */
-#include "cfe_evs_utils.h"  /* EVS utility function definitions */
-#include "cfe_evs_dispatch.h"
-
-#endif /* CFE_EVS_MODULE_ALL_H */
+#endif /* CFE_SB_DISPATCH_H */
