+++ modules/sb/fsw/src/cfe_sb_module_all.h
@@ -43,6 +43,7 @@
 #include "cfe_sb_events.h"
 #include "cfe_sb_destination_typedef.h"
 #include "cfe_sb_msg.h"
+#include "cfe_sb_dispatch.h"
 #include "cfe_sbr.h"
 #include "cfe_core_resourceid_basevalues.h"
