+++ modules/tbl/CMakeLists.txt
@@ -12,10 +12,7 @@ set(tbl_SOURCES
     fsw/src/cfe_tbl_internal.c
     fsw/src/cfe_tbl_task.c
     fsw/src/cfe_tbl_task_cmds.c
-    fsw/src/cfe_tbl_api.c
-    fsw/src/cfe_tbl_internal.c
-    fsw/src/cfe_tbl_task.c
-    fsw/src/cfe_tbl_task_cmds.c
+    fsw/src/cfe_tbl_dispatch.c
 )
 add_library(tbl STATIC ${tbl_SOURCES})
