+++ modules/tbl/fsw/src/cfe_tbl_module_all.h
@@ -44,6 +44,7 @@
 #include "cfe_tbl_internal.h"
 #include "cfe_tbl_task.h"
 #include "cfe_tbl_task_cmds.h"
+#include "cfe_tbl_dispatch.h"
 
 /*
  * Additionally TBL needs to use special/extra CDS APIs that are not in the normal API
