+++ modules/time/fsw/src/cfe_time_dispatch.h
@@ -16,33 +16,32 @@
  * limitations under the License.
  ************************************************************************/
 
-/**
- * @file
+/*
+** File: cfe_time_task.c
+**
+** Subsystem: cFE TIME Task
+**
+** Author: S. Walling (Microtel)
+**
+** Notes:
+**
+*/
+#ifndef CFE_TIME_DISPATCH_H
+#define CFE_TIME_DISPATCH_H
+
+/*
+** Required header files...
+*/
+#include "cfe_time_api_typedefs.h"
+#include "cfe_sb_api_typedefs.h"
+#include "cfe_msg_api_typedefs.h"
+
+/*----------------------------------------------------------------
  *
- * Encapsulates all EVS module internal header files, as well
- * as the public API from all other CFE core modules, OSAL, and PSP.
+ * Application-scope internal function
+ * See description in header file for argument/return detail
  *
- * This simplifies the set of include files that need to be put at the
- * start of every source file.
- */
+ *-----------------------------------------------------------------*/
+void CFE_TIME_TaskPipe(const CFE_SB_Buffer_t *SBBufPtr);
 
-#ifndef CFE_EVS_MODULE_ALL_H
-#define CFE_EVS_MODULE_ALL_H
-
-/********************* Include Files  ************************/
-
-#include "cfe.h" /* All CFE+OSAL public API definitions */
-#include "cfe_platform_cfg.h"
-
-#include "cfe_msgids.h"
-#include "cfe_perfids.h"
-
-#include "cfe_evs_core_internal.h"
-
-#include "cfe_evs_events.h" /* EVS event IDs */
-#include "cfe_evs_task.h"   /* EVS internal definitions */
-#include "cfe_evs_log.h"    /* EVS log file definitions */
-#include "cfe_evs_utils.h"  /* EVS utility function definitions */
-#include "cfe_evs_dispatch.h"
-
-#endif /* CFE_EVS_MODULE_ALL_H */
+#endif
