+++ modules/time/fsw/src/cfe_time_module_all.h
@@ -42,5 +42,6 @@
 #include "cfe_time_msg.h"
 #include "cfe_time_events.h"
 #include "cfe_time_utils.h"
+#include "cfe_time_dispatch.h"
 
 #endif /* CFE_TIME_MODULE_ALL_H */
