+++ modules/time/fsw/src/cfe_time_task.c
@@ -307,242 +307,6 @@ int32 CFE_TIME_TaskInit(void)
 
 /*----------------------------------------------------------------
  *
- * Internal helper routine only, not part of API.
- *
- * Function to verify the length of incoming TIME command packets
- *
- *-----------------------------------------------------------------*/
-bool CFE_TIME_VerifyCmdLength(CFE_MSG_Message_t *MsgPtr, size_t ExpectedLength)
-{
-    bool              result       = true;
-    CFE_MSG_Size_t    ActualLength = 0;
-    CFE_MSG_FcnCode_t FcnCode      = 0;
-    CFE_SB_MsgId_t    MsgId        = CFE_SB_INVALID_MSG_ID;
-
-    CFE_MSG_GetSize(MsgPtr, &ActualLength);
-
-    /*
-    ** Verify the command packet length
-    */
-    if (ExpectedLength != ActualLength)
-    {
-        CFE_MSG_GetMsgId(MsgPtr, &MsgId);
-        CFE_MSG_GetFcnCode(MsgPtr, &FcnCode);
-
-        CFE_EVS_SendEvent(CFE_TIME_LEN_ERR_EID, CFE_EVS_EventType_ERROR,
-                          "Invalid msg length: ID = 0x%X,  CC = %u, Len = %u, Expected = %u",
-                          (unsigned int)CFE_SB_MsgIdToValue(MsgId), (unsigned int)FcnCode, (unsigned int)ActualLength,
-                          (unsigned int)ExpectedLength);
-        result = false;
-        ++CFE_TIME_Global.CommandErrorCounter;
-    }
-
-    return result;
-}
-
-/*----------------------------------------------------------------
- *
- * Application-scope internal function
- * See description in header file for argument/return detail
- *
- *-----------------------------------------------------------------*/
-void CFE_TIME_TaskPipe(CFE_SB_Buffer_t *SBBufPtr)
-{
-    CFE_SB_MsgId_t    MessageID   = CFE_SB_INVALID_MSG_ID;
-    CFE_MSG_FcnCode_t CommandCode = 0;
-
-    CFE_MSG_GetMsgId(&SBBufPtr->Msg, &MessageID);
-
-    switch (CFE_SB_MsgIdToValue(MessageID))
-    {
-        /*
-        ** Housekeeping telemetry request...
-        */
-        case CFE_TIME_SEND_HK_MID:
-            CFE_TIME_HousekeepingCmd((CFE_TIME_SendHkCmd_t *)SBBufPtr);
-            break;
-
-        /*
-        ** Time at the tone "signal"...
-        */
-        case CFE_TIME_TONE_CMD_MID:
-            CFE_TIME_ToneSignalCmd((CFE_TIME_ToneSignalCmd_t *)SBBufPtr);
-            break;
-
-        /*
-        ** Time at the tone "data"...
-        */
-        case CFE_TIME_DATA_CMD_MID:
-            CFE_TIME_ToneDataCmd((CFE_TIME_ToneDataCmd_t *)SBBufPtr);
-            break;
-
-        /*
-        ** Run time state machine at 1Hz...
-        */
-        case CFE_TIME_1HZ_CMD_MID:
-            CFE_TIME_OneHzCmd((CFE_TIME_1HzCmd_t *)SBBufPtr);
-            break;
-
-/*
-** Request for time at the tone "data"...
-*/
-#if (CFE_PLATFORM_TIME_CFG_SERVER == true)
-        case CFE_TIME_SEND_CMD_MID:
-            CFE_TIME_ToneSendCmd((CFE_TIME_FakeToneCmd_t *)SBBufPtr);
-            break;
-#endif
-
-        /*
-        ** Time task ground commands...
-        */
-        case CFE_TIME_CMD_MID:
-
-            CFE_MSG_GetFcnCode(&SBBufPtr->Msg, &CommandCode);
-            switch (CommandCode)
-            {
-                case CFE_TIME_NOOP_CC:
-                    if (CFE_TIME_VerifyCmdLength(&SBBufPtr->Msg, sizeof(CFE_TIME_NoopCmd_t)))
-                    {
-                        CFE_TIME_NoopCmd((CFE_TIME_NoopCmd_t *)SBBufPtr);
-                    }
-                    break;
-
-                case CFE_TIME_RESET_COUNTERS_CC:
-                    if (CFE_TIME_VerifyCmdLength(&SBBufPtr->Msg, sizeof(CFE_TIME_ResetCountersCmd_t)))
-                    {
-                        CFE_TIME_ResetCountersCmd((CFE_TIME_ResetCountersCmd_t *)SBBufPtr);
-                    }
-                    break;
-
-                case CFE_TIME_SEND_DIAGNOSTIC_TLM_CC:
-                    if (CFE_TIME_VerifyCmdLength(&SBBufPtr->Msg, sizeof(CFE_TIME_SendDiagnosticCmd_t)))
-                    {
-                        CFE_TIME_SendDiagnosticTlm((CFE_TIME_SendDiagnosticCmd_t *)SBBufPtr);
-                    }
-                    break;
-
-                case CFE_TIME_SET_STATE_CC:
-                    if (CFE_TIME_VerifyCmdLength(&SBBufPtr->Msg, sizeof(CFE_TIME_SetStateCmd_t)))
-                    {
-                        CFE_TIME_SetStateCmd((CFE_TIME_SetStateCmd_t *)SBBufPtr);
-                    }
-                    break;
-
-                case CFE_TIME_SET_SOURCE_CC:
-                    if (CFE_TIME_VerifyCmdLength(&SBBufPtr->Msg, sizeof(CFE_TIME_SetSourceCmd_t)))
-                    {
-                        CFE_TIME_SetSourceCmd((CFE_TIME_SetSourceCmd_t *)SBBufPtr);
-                    }
-                    break;
-
-                case CFE_TIME_SET_SIGNAL_CC:
-                    if (CFE_TIME_VerifyCmdLength(&SBBufPtr->Msg, sizeof(CFE_TIME_SetSignalCmd_t)))
-                    {
-                        CFE_TIME_SetSignalCmd((CFE_TIME_SetSignalCmd_t *)SBBufPtr);
-                    }
-                    break;
-
-                /*
-                ** Time Clients process "tone delay" commands...
-                */
-                case CFE_TIME_ADD_DELAY_CC:
-                    if (CFE_TIME_VerifyCmdLength(&SBBufPtr->Msg, sizeof(CFE_TIME_AddDelayCmd_t)))
-                    {
-                        CFE_TIME_AddDelayCmd((CFE_TIME_AddDelayCmd_t *)SBBufPtr);
-                    }
-                    break;
-
-                case CFE_TIME_SUB_DELAY_CC:
-                    if (CFE_TIME_VerifyCmdLength(&SBBufPtr->Msg, sizeof(CFE_TIME_SubDelayCmd_t)))
-                    {
-                        CFE_TIME_SubDelayCmd((CFE_TIME_SubDelayCmd_t *)SBBufPtr);
-                    }
-                    break;
-
-                /*
-                ** Time Servers process "set time" commands...
-                */
-                case CFE_TIME_SET_TIME_CC:
-                    if (CFE_TIME_VerifyCmdLength(&SBBufPtr->Msg, sizeof(CFE_TIME_SetTimeCmd_t)))
-                    {
-                        CFE_TIME_SetTimeCmd((CFE_TIME_SetTimeCmd_t *)SBBufPtr);
-                    }
-                    break;
-
-                case CFE_TIME_SET_MET_CC:
-                    if (CFE_TIME_VerifyCmdLength(&SBBufPtr->Msg, sizeof(CFE_TIME_SetMETCmd_t)))
-                    {
-                        CFE_TIME_SetMETCmd((CFE_TIME_SetMETCmd_t *)SBBufPtr);
-                    }
-                    break;
-
-                case CFE_TIME_SET_STCF_CC:
-                    if (CFE_TIME_VerifyCmdLength(&SBBufPtr->Msg, sizeof(CFE_TIME_SetSTCFCmd_t)))
-                    {
-                        CFE_TIME_SetSTCFCmd((CFE_TIME_SetSTCFCmd_t *)SBBufPtr);
-                    }
-                    break;
-
-                case CFE_TIME_SET_LEAP_SECONDS_CC:
-                    if (CFE_TIME_VerifyCmdLength(&SBBufPtr->Msg, sizeof(CFE_TIME_SetLeapSecondsCmd_t)))
-                    {
-                        CFE_TIME_SetLeapSecondsCmd((CFE_TIME_SetLeapSecondsCmd_t *)SBBufPtr);
-                    }
-                    break;
-
-                case CFE_TIME_ADD_ADJUST_CC:
-                    if (CFE_TIME_VerifyCmdLength(&SBBufPtr->Msg, sizeof(CFE_TIME_AddAdjustCmd_t)))
-                    {
-                        CFE_TIME_AddAdjustCmd((CFE_TIME_AddAdjustCmd_t *)SBBufPtr);
-                    }
-                    break;
-
-                case CFE_TIME_SUB_ADJUST_CC:
-                    if (CFE_TIME_VerifyCmdLength(&SBBufPtr->Msg, sizeof(CFE_TIME_SubAdjustCmd_t)))
-                    {
-                        CFE_TIME_SubAdjustCmd((CFE_TIME_SubAdjustCmd_t *)SBBufPtr);
-                    }
-                    break;
-
-                case CFE_TIME_ADD_1HZ_ADJUSTMENT_CC:
-                    if (CFE_TIME_VerifyCmdLength(&SBBufPtr->Msg, sizeof(CFE_TIME_Add1HZAdjustmentCmd_t)))
-                    {
-                        CFE_TIME_Add1HZAdjustmentCmd((CFE_TIME_Add1HZAdjustmentCmd_t *)SBBufPtr);
-                    }
-                    break;
-
-                case CFE_TIME_SUB_1HZ_ADJUSTMENT_CC:
-                    if (CFE_TIME_VerifyCmdLength(&SBBufPtr->Msg, sizeof(CFE_TIME_Sub1HZAdjustmentCmd_t)))
-                    {
-                        CFE_TIME_Sub1HZAdjustmentCmd((CFE_TIME_Sub1HZAdjustmentCmd_t *)SBBufPtr);
-                    }
-                    break;
-
-                default:
-
-                    CFE_TIME_Global.CommandErrorCounter++;
-                    CFE_EVS_SendEvent(CFE_TIME_CC_ERR_EID, CFE_EVS_EventType_ERROR,
-                                      "Invalid command code -- ID = 0x%X, CC = %d",
-                                      (unsigned int)CFE_SB_MsgIdToValue(MessageID), (int)CommandCode);
-                    break;
-            } /* switch (CFE_TIME_CMD_MID -- command code)*/
-            break;
-
-        default:
-
-            /*
-            ** Note: we only increment the command error counter when
-            **    processing CFE_TIME_CMD_MID commands...
-            */
-            CFE_EVS_SendEvent(CFE_TIME_ID_ERR_EID, CFE_EVS_EventType_ERROR, "Invalid message ID -- ID = 0x%X",
-                              (unsigned int)CFE_SB_MsgIdToValue(MessageID));
-            break;
-
-    } /* switch (message ID) */
-}
-
-/*----------------------------------------------------------------
- *
  * Application-scope internal function
  * See description in header file for argument/return detail
  *
